CHAT_BOT/phase2/pyproject.toml
[tool.poetry]
name = "phase2"
version = "0.1.0"
description = ""
authors = ["expectolimited <devpatel232408@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
promptlayer = "^1.0.6"
openai = "^1.35.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
